+++ src/AppCore.h
@@ -17,8 +17,8 @@ private:
 
 private:
     void _printHelp() noexcept;
-    void _addTrainings(const uint16_t num) noexcept;
-    void _setTrainings(const uint16_t num) noexcept;
+    void _addTrainings(const uint32_t num) noexcept;
+    void _setTrainings(const uint32_t num) noexcept;
     void _markTraining() noexcept;
     void _showTrainings() noexcept;
     void _removeLogfile();
